Fixed exception viewing profile
Fixed an issue where when a user who does not have any activity in the database tried to view their profile, would be greeted with an exception error. The fix was a change to the Calculate controller so that null was checked for properly before the array was accessed, and not the other way around.