Basic Api Support
This commit adds the beginnings of api support for the application. At the moment, hashed api tokens are stored in the user table, but are null by default (as they are one-way hashed). For now, to get a token you go to http://127.0.0.1:8000/api/gimme and that will spit out an unhashed token, and store the hashed token in your row on the user table (as such, you need to be signed in to get a token). The token can then be used in post requests to endpoints at the version 1 url (http://127.0.0.1:8000/api/v1). At the moment, the only endpoint is http://127.0.0.1:8000/api/v1/info.

Tokens must be passed in the header of the POST request, using the 'Authorization' key, with a value of 'Bearer <token>'. I created a super simple documentation site for the api, found here: https://antelope.docs.apiary.io/, which I will built upon as I go.

I've created two api controllers, one that handles token creation, and in the future will be used to handle creation, deletion, etc; the second handles the actual endpoints. I'm not sure if this is the best way to go about this, so I'm open to suggestions.

I've also created two api resources, one handles the successful responses, the other handles the failed responses. The only real difference is the value of the 'status' response, as I could not figure out how to combined them in one file, someone feel free to enlighten me. HTTP status codes are set per-return.