$_DEV: Change user password from member settings
This commit allows admins to change a user's password to a temp password of their choosing. Temp passwords must be at least 3 characters long.

This commit adds a `editpassword` method to the  Edit Profile controller; adds custom css to the custom.css file to override css injected by js by sweetalerts; adds a sweetalert to the js file for the edit profile model; adds a new route for the previously mentioned controller.