$_DEV: Edit flags button and modal
This commit adds a 'Flags' button next to patrol logs in the activity database, clicking this button opens a new Flags modal that populates with the flag data from that patrol. If there is no flag, self or auto, the button to resolve the flag is disabled, and cannot be used. If there is a flag, the staff member can click the respective checkbox, which then opens up an optional textbox to add a note, then they can click submit. The submit button (and associated save controller) had not yet been written; it's all for looks at the moment.