$_DEV: Dashboard reloads data on patrol log submit
Kinda hacky, but when you submit the PAL form, it includes a hidden input telling the JS what page you're on, and if you're on the dashboard, it reloads just the middle section of the page using jQuery, so the PHP loads as if it was a refresh. No validation since it's not really needed.